Port C neighbour search algorithm to Rcpp-C++ maintaining old sources for testing.
The old C-code is maintained to compare the ported version with the old one
(see test_neighbour_search_port.R) and to allow the old tests to be succesful
(corrDim, infDim, etc.). To compare the new versions with the old ones, new
R functions are created by prepending the rcpp prefix to the old names. However,
the shall be eliminated at the end of the code port.